ENH: Add source loc to Report
